+++ RPi/GPIO_DEVEL/__init__.py
@@ -1,3 +1,21 @@
-# Development functions, not needed for normal use
-from RPi._GPIO import Reset, State_Access, setdebuginfo, is_all_ints, is_all_bools_or_directions,\
-    is_iterable, line_get_mode, line_set_mode, _line_mode_none, _line_mode_out, _line_mode_in
+"""
+The new RPi.GPIO_DEVEL development and debug API
+By Joel Savitz and Fabrizio D'Angelo
+This is free software, see LICENSE for details
+"""
+
+# We have added functions and constants to this list as we have seen
+# necesary but we are open to adding more if there is any interest
+
+from RPi.core import\
+    is_all_bools_or_directions,\
+    is_all_ints,\
+    is_iterable,\
+    Reset,\
+    setdebuginfo,\
+    State_Access,\
+    line_get_mode,\
+    line_set_mode,\
+    _line_mode_in,\
+    _line_mode_none,\
+    _line_mode_out
