+++ spec/sources.bib
@@ -1,17 +1,17 @@
 @misc{rpigpio,
   title = {RPi.GPIO PyPi project page},
-  howpublished = {\url{https://pypi.org/project/RPi.GPIO/}},
+  howpublished = {\texttt{https://pypi.org/project/RPi.GPIO/}},
   note = {Accessed: 2020-05-14}
 }
 
 @misc{libgpiod,
   title = {libgpiod mainline repository},
-  howpublished = {\url{https://git.kernel.org/pub/scm/libs/libgpiod/libgpiod.git/}},
+  howpublished = {\texttt{https://git.kernel.org/pub/scm/libs/libgpiod/libgpiod.git/}},
   note = {Accessed: 2020-05-14}
 }
 
 @misc{gpiozero,
   title = {gpiozero mainline repository},
-  howpublished = {\url{https://github.com/gpiozero/gpiozero}},
+  howpublished = {\texttt{https://github.com/gpiozero/gpiozero}},
   note = {Accessed: 2020-05-19}
 }
